Tests are combined into one file
