+++ homework/homework1/count_the_equation.py
@@ -5,7 +5,7 @@ import random
 from math import e, sin, radians
 
 
-def equation(a: int, x: int) -> tuple[float, float]:
+def equation1(a: int, x: int) -> tuple[float, float]:
     if not -20 <= a <= 20 or not 0 <= x <= 100:
         raise ValueError("out of range")
     result_radians = e**a * sin(x) ** 2 - abs(x - a) / 7
@@ -21,7 +21,7 @@ if __name__ == "__main__":
     print(f"a = {a1}")
     print(f"x = {x1}")
 
-    y_radians, y_degrees = equation(a1, x1)
+    y_radians, y_degrees = equation1(a1, x1)
 
     print(f"If x is given in radians, у = {y_radians:.4f}")
     print(f"If x is given in degrees, y = {y_degrees:.4f}")
