+++ homework/homework1/perimeter_and_area_of_a_triangle.py
@@ -22,7 +22,7 @@ def distance(d1: tuple[int, int], d2: tuple[int, int]) -> float:
     return round(math.sqrt((d2[0] - d1[0]) ** 2 + (d2[1] - d1[1]) ** 2), 2)
 
 
-def calculations(a: tuple[int, int], b: tuple[int, int], c: tuple[int, int]) -> float:
+def calculation_perimetr(a: tuple[int, int], b: tuple[int, int], c: tuple[int, int]) -> float:
     return round(distance(b, c) + distance(a, c) + distance(a, b), 2)
 
 
@@ -40,7 +40,7 @@ if __name__ == "__main__":
     print(distance(a1, c1))
     print(distance(a1, b1))
 
-    result_perimetr = calculations(a1, b1, c1)
+    result_perimetr = calculation_perimetr(a1, b1, c1)
     result_area = area_of_triangle(a1, b1, c1)
 
     print(f"The perimeter of triangle ABC is: {result_perimetr:.2f}")
