+++ homework/homework1/the_meaning_of_the_expression.py
@@ -4,7 +4,7 @@ Calculate the value of the expression y = 3ab - 4 / (c * d)"""
 import random
 
 
-def calculation(a: int, b: int, c: int, d: int) -> float:
+def calculation_expression(a: int, b: int, c: int, d: int) -> float:
     if not all(-100 <= x <= 100 for x in (a, b, c, d)):
         raise ValueError("out of range")
     return round(3 * a * b - 4 / (c * d), 2)
@@ -22,6 +22,6 @@ if __name__ == "__main__":
     print(f"integer c = {c1}")
     print(f"integer d = {d1}")
 
-    y = calculation(a1, b1, c1, d1)
+    y = calculation_expression(a1, b1, c1, d1)
 
     print(f'Result of the expression "y = 3ab - 4 / (c * d)": {y:.2f}')
