+++ homework/homework1/the_sum_of_the_numbers.py
@@ -4,7 +4,7 @@ find the sum of its digits"""
 import random
 
 
-def calculation(x: int) -> int:
+def calculation_sum_num(x: int) -> int:
     if not 100 <= x <= 999:
         raise ValueError("out of range")
     return x // 100 + x % 100 // 10 + x % 10
@@ -16,6 +16,6 @@ if __name__ == "__main__":
 
     print(f"the three-digit number = {num}")
 
-    result = calculation(num)
+    result = calculation_sum_num(num)
 
     print(f"Sum of the digits of a number {num}, equal: {result}")
